+++ src/GUI/SelectFontUI.cpp
@@ -4,6 +4,7 @@
 #include "../Localisation/LocalisationManager.hpp"
 #include "BetterAlertLayer.hpp"
 #include <Button.hpp>
+#include <ScrollLayer.hpp>
 
 using namespace qolmod;
 
@@ -267,20 +268,16 @@ bool SelectFontUI::setup()
     scrollbar2 = BetterScrollbar::create(scroll2);
     scrollbar2->setDisabled(true);
 
-    scroll1 = ScrollLayer::create(ccp(280 - 15, 160));
-    scroll1->setUserObject("MouseDispatcherExt"_spr, CCNode::create());
+    scroll1 = qolmod::ScrollLayer::create(ccp(280 - 15, 160));
     scroll1->m_contentLayer->setContentHeight(height);
     scroll1->m_peekLimitTop = 15;
     scroll1->m_peekLimitBottom = 15;
-    scroll1->setUserObject("user95401.scrollbar_everywhere/scrollbar", CCNode::create());
 
-    scroll2 = ScrollLayer::create(ccp(280 - 15, 160));
-    scroll2->setUserObject("MouseDispatcherExt"_spr, CCNode::create());
+    scroll2 = qolmod::ScrollLayer::create(ccp(280 - 15, 160));
     scroll2->m_peekLimitTop = 15;
     scroll2->m_peekLimitBottom = 15;
     scroll2->setMouseEnabled(false);
     scroll2->setTouchEnabled(false);
-    scroll2->setUserObject("user95401.scrollbar_everywhere/scrollbar", CCNode::create());
 
     scrollbar1->setLayer(scroll1);
     scrollbar2->setLayer(scroll2);